requirements.txt
Flask==1.0.2
Flask-MySQL==1.4.0
passlib==1.7.1
Flask-MySQLdb==0.2.0
requests==2.20.0
